Move solver to separate package
